[refactor] Create Otamatone Module File
